services_1/ml_pipeline/requirements.txt
pandas==1.3.1
scikit-learn==0.24.2
joblib==1.0.1
psycopg2-binary==2.9.1